Make autoload
Move functions to autoload. Increment version to 2.0

new_buffer is the only function used in the API, so all the rest can
stay as script-local functions.

s:unscratch is called from a command, but I confirmed that it still
works (the command is defined within a function in the same file).